WIP: Adding fuzz targets for oss-fuzz
The strategy is to replace the read(2) function for reading socket data
with a mock function that returns the fuzzed data. Then a HTTP session is
created and the http_session() function is called. The mock function can
be found in fuzz_util.c. The mock has to return -1 (socket error) when
all data has been delivered, presumably because this is how the real
read(2) API signals that the client has completed sending the request.

This is how the read(2) API signals that a HTTP client has disconnected
after delivering it's request.

Added fuzz/ folder for fuzzing endpoint and related code.
fuzz/request.corpus contains the seed corpus of requests that will be
mutated during fuzz testing.

The fuzz/test/ folder contains unit tests for the fuzzing glue code itself.

Added CMake build system for better header dependency detection and ctest runner.